+++ NEWS.md
@@ -10,6 +10,10 @@
 
 * **Dropped support for running on Node &lt; 4.0.** (#319)
 
+### Fixed
+
+* `strict-ssl` (and by extension, `download.strictSSL`) defaults to `true`, as documented (#320)
+
 ### Deprecated
 
 * `cache` is deprecated in favor of `download.cache` (#320)